+++ ios/ScreenTimeModule.swift
@@ -41,7 +41,11 @@ class ScreenTimeModule: NSObject {
     return container.mainContext
   }
   
-  func handleAuthorizationError(_ errorCode: String? = nil, error: Error, reject: @escaping RCTPromiseRejectBlock) {
+  func handleAuthorizationError(
+    _ errorCode: String? = nil,
+    error: Error,
+    reject: @escaping RCTPromiseRejectBlock
+  ) {
       let finalErrorCode = errorCode ?? "FAILED"
       let errorMessage = "Failed to request authorization: \(error.localizedDescription)"
       reject(finalErrorCode, errorMessage, error)
@@ -65,7 +69,11 @@ class ScreenTimeModule: NSObject {
   }
   
   @MainActor @objc
-  func requestAuthorization(_ testBlockName: String = "", resolve: @escaping RCTPromiseResolveBlock, rejecter reject: @escaping RCTPromiseRejectBlock) {
+  func requestAuthorization(
+    _ testBlockName: String = "",
+    resolve: @escaping RCTPromiseResolveBlock,
+    rejecter reject: @escaping RCTPromiseRejectBlock
+  ) {
     if #available(iOS 16.0, *) {
       Task {
         do {
@@ -96,7 +104,15 @@ class ScreenTimeModule: NSObject {
   }
 
   @MainActor @objc
-  func showAppPicker(_ isFirstSelection: Bool, blockId: String = "", limitId: String = "", saveButtonText: String = "", titleText: String = "", resolve: @escaping RCTPromiseResolveBlock, rejecter reject: @escaping RCTPromiseRejectBlock) {
+  func showAppPicker(
+    _ isFirstSelection: Bool,
+    blockId: String = "",
+    limitId: String = "",
+    saveButtonText: String = "",
+    titleText: String = "",
+    resolve: @escaping RCTPromiseResolveBlock,
+    rejecter reject: @escaping RCTPromiseRejectBlock
+  ) {
     DispatchQueue.main.async {
       
       if let scene = UIApplication.shared.connectedScenes.first as? UIWindowScene,
@@ -134,7 +150,14 @@ class ScreenTimeModule: NSObject {
   }
 
   @MainActor @objc
-  func createBlock(_ name: String, startTime: String, endTime: String, weekdays: [Int], resolve: @escaping RCTPromiseResolveBlock, rejecter reject: @escaping RCTPromiseRejectBlock) {
+  func createBlock(
+    _ name: String,
+    startTime: String,
+    endTime: String,
+    weekdays: [Int],
+    resolve: @escaping RCTPromiseResolveBlock,
+    rejecter reject: @escaping RCTPromiseRejectBlock
+  ) {
     let deviceActivityCenter = DeviceActivityCenter();
     
     do {
@@ -186,17 +209,6 @@ class ScreenTimeModule: NSObject {
     }
   }
   
-  func getLimitTime(time: String = "") -> Int? {
-    let components = time.split(separator: ":")
-    if let hours = Int(components[0]), let minutes = Int(components[1]) {
-      let totalMinutes = (hours * 60) + minutes
-      return totalMinutes
-    } else {
-      print("Invalid format")
-      return nil
-    }
-  }
-  
   @MainActor @objc
   func createLimit(
     _ name: String, 
@@ -234,7 +246,7 @@ class ScreenTimeModule: NSObject {
       // Create event for each app or web
       var eventsArray: [DeviceActivityEvent.Name: DeviceActivityEvent] = [:]
       
-      let minutesToBlock = getLimitTime(time: timeLimit);
+      let minutesToBlock = LimitModule.shared.getLimitTime(time: timeLimit);
       
       // Create events
       try self.appsSelected.forEach{appSelected in
@@ -327,275 +339,6 @@ class ScreenTimeModule: NSObject {
     }
   }
   
-  
-  @MainActor
-  func findLimits() -> [Limit]{
-    
-    do {
-      let context = try getContext()
-      
-      let fetchDescriptor = FetchDescriptor<Limit>()
-      let limits = try context.fetch(fetchDescriptor)
-      
-      return limits
-    
-    } catch {
-      print("Error trying to get limits")
-      return []
-    }
-  }
-  
-  @MainActor @objc
-  func getLimits(_ impulseMode: Bool = false, resolve: @escaping RCTPromiseResolveBlock, rejecter reject: @escaping RCTPromiseRejectBlock){
-    do {
-      let limits = findLimits()
-      
-      let limitArray = limits.map { limit -> [String: Any] in
-        return [
-            "id": limit.id.uuidString, // Asegúrate de que 'id' sea un UUID
-            "title": limit.name, // Reemplaza con los campos de tu modelo
-            "timeLimit": limit.timeLimit,
-            "openLimit": limit.openLimit,
-            "apps": limit.appsTokens.count,
-            "weekdays": limit.weekdays,
-            "enable": limit.enable
-        ]
-      }
-      resolve(["status": "success", "limits" : limitArray])
-    } catch {
-      print("Error finding limits")
-    }
-  }
-  
-  
-  @MainActor
-  func disableLimit(limitId: UUID, updateStore: Bool = false){
-    do {
-      // Remove shields
-      let limit = findLimit(limitId: limitId)
-      limit?.appsEvents.forEach{event in
-        let store = ManagedSettingsStore(named: ManagedSettingsStore.Name(rawValue: "event-\(event.id.uuidString)"))
-        store.shield.applications = nil
-      }
-      
-      // Stop monitoring
-      let deviceActivityCenter = DeviceActivityCenter();
-
-      // Validate if weekdays is upper 0
-      if limit?.weekdays.count ?? 0 > 0 {
-        // Remove for each day
-        limit?.weekdays.forEach { weekday in
-          let deviceActivityCenter = DeviceActivityCenter();
-          deviceActivityCenter.stopMonitoring([DeviceActivityName(rawValue: "\(limitId.uuidString)-limit-day-\(weekday)")])
-        }
-      } else {
-        deviceActivityCenter.stopMonitoring([DeviceActivityName(rawValue: "\(String(describing: limitId.uuidString))-limit")])
-      }
-      
-      // Update limit status
-      if updateStore {
-        limit?.enable = false
-        try limit?.modelContext?.save()
-      }
-      
-    } catch {
-      print("Error trying to disable limit")
-    }
-  }
-  
-  @MainActor
-  func enableLimit(limitId: UUID, updateStore: Bool = false){
-    do {
-      
-      // Get limit
-      let limit = findLimit(limitId: limitId)
-      
-      // Get minutes to block for each event
-      let minutesToBlock = getLimitTime(time: limit?.timeLimit ?? "");
-      
-      var eventsArray: [DeviceActivityEvent.Name: DeviceActivityEvent] = [:]
-      limit?.appsEvents.forEach{event in
-        if minutesToBlock != nil {
-          let threshold = DateComponents(minute: minutesToBlock)
-          let eventName = DeviceActivityEvent.Name(rawValue: "\(event.id.uuidString)-event")
-          let activityEvent = DeviceActivityEvent(applications: [event.appToken], threshold: threshold)
-                    
-          eventsArray[eventName] = activityEvent
-          
-          // Find if the event has history
-          let numberOfEvents = findLimitHistory(event: event)
-          print("Number of events: \(numberOfEvents)")
-          if numberOfEvents! > 0 {
-            // Block app if the event has history
-            let store = ManagedSettingsStore(named: ManagedSettingsStore.Name(rawValue: "event-\(event.id.uuidString)"))
-            store.shield.applications = [event.appToken]
-          }
-          
-        }
-      }
-      
-      // Start monitoring
-      let deviceActivityCenter = DeviceActivityCenter();
-      
-      let activities = deviceActivityCenter.activities
-      
-      print("current activities \(activities)")
-      
-      let weekdays: [Int] = limit?.weekdays ?? []
-      
-      // Validate if weekdays is upper 0
-      if weekdays.count > 0 {
-        logger.info("Enable frecuency")
-        try weekdays.forEach { weekday in
-          logger.info("Create monitoring with weekday: \(weekday)")
-          try deviceActivityCenter.startMonitoring(
-            DeviceActivityName(rawValue: "\(limit?.id.uuidString)-limit-day-\(weekday)"),
-            during: DeviceActivitySchedule(
-              intervalStart: DateComponents(hour: 0, minute: 0, weekday: weekday),
-              intervalEnd: DateComponents(hour: 23, minute: 59, weekday: weekday),
-              repeats: true
-            ),
-            events: eventsArray
-          )
-        }
-      } else {
-        try deviceActivityCenter.startMonitoring(
-          DeviceActivityName(rawValue: "\(limitId.uuidString)-limit"),
-          during: DeviceActivitySchedule(
-            intervalStart: DateComponents(hour: 0, minute: 0),
-            intervalEnd: DateComponents(hour: 23, minute: 59),
-            repeats: false
-          ),
-          events: eventsArray
-        )
-      }
-
-      // Update limit status
-      if updateStore {
-        limit?.enable = true
-        try limit?.modelContext?.save()
-      }
-    } catch {
-      print("Error trying to enable limit")
-    }
-  }
-  
-  @MainActor
-  func findLimitHistory(event: AppEvent) -> Int? {
-    do {
-      
-      let calendar = Calendar.current
-      let startOfDay = calendar.startOfDay(for: Date())
-      let endOfDay = calendar.date(byAdding: .day, value: 1, to: startOfDay)!
-      
-      let history = try event.history.filter(#Predicate{ $0.date >= startOfDay && $0.date < endOfDay })
-      
-      return history.count
-    } catch {
-      print("Error trying to get limit history")
-      return nil
-    }
-  }
-  
-  @MainActor
-  func findLimit(limitId: UUID) -> Limit? {
-    do {
-      let context = try getContext()
-      
-      // Find limit
-      let fetchDescriptor = FetchDescriptor<Limit>(
-        predicate: #Predicate{ $0.id == limitId }
-      )
-      let limit = try context.fetch(fetchDescriptor)
-      return limit.first
-    } catch {
-      print("Error trying to find limit with events")
-      return nil
-    }
-  }
-  
-  @MainActor @objc
-  func deleteLimit(_ limitId: String = "", resolve: @escaping RCTPromiseResolveBlock, rejecter reject: @escaping RCTPromiseRejectBlock){
-    do {
-      guard let uuid = UUID(uuidString: limitId) else {
-        reject("invalid_uuid", "The limit id is not a valid UUID", nil)
-        return
-      }
-      
-      // Get events
-      let limit = findLimit(limitId: uuid)
-      let context = try getContext()
-      
-      limit?.appsEvents.forEach{event in
-        // Remove restrictions for every app
-        let store = ManagedSettingsStore(named: ManagedSettingsStore.Name(rawValue: "event-\(event.id)"))
-        store.shield.applications = nil
-      }
-      
-      // Stop monitoring
-      let deviceActivityCenter = DeviceActivityCenter();
-
-      // Validate if weekdays is upper 0
-      if (limit?.weekdays.count)! > 0 {
-        // Remove for each day
-        limit?.weekdays.forEach { weekday in
-          deviceActivityCenter.stopMonitoring([DeviceActivityName(rawValue: "\(uuid)-limit-day-\(weekday)")])
-        }
-      } else {
-        deviceActivityCenter.stopMonitoring([DeviceActivityName(rawValue: "\(uuid)-limit")])
-      }
-            
-      // Delete limit and events
-      try context.delete(model: Limit.self, where: #Predicate { $0.id == uuid })
-      resolve(["status": "success"])
-    } catch {
-      print("Error trying to delete limit")
-    }
-  }
-  
-  @MainActor @objc
-  func updateLimitStatus(_ limitId: String = "", enable: Bool, resolve: @escaping RCTPromiseResolveBlock, rejecter reject: @escaping RCTPromiseRejectBlock) {
-    guard let uuid = UUID(uuidString: limitId) else {
-      reject("invalid_uuid", "The limit id is not a valid UUID", nil)
-      return
-    }
-    
-    if !enable {
-      // Disable limint and remove events shield
-      print("Disable limit")
-      disableLimit(limitId: uuid, updateStore: true)
-    } else {
-      // Enable limit and events again
-      print("Enable limit")
-      enableLimit(limitId: uuid, updateStore: true)
-    }
-    
-    resolve(["status": "success"])
-  }
-  
-  @MainActor @objc
-  func getLimitDetail(_ limitId: String, resolve: @escaping RCTPromiseResolveBlock, rejecter reject: @escaping RCTPromiseRejectBlock){
-    guard let uuid = UUID(uuidString: limitId) else {
-      reject("invalid_uuid", "The limit id is not a valid UUID", nil)
-      return
-    }
-    
-    let limit = findLimit(limitId: uuid)
-    
-    let limitData = [
-      "id": limit?.id.uuidString,
-      "name": limit?.name,
-      "timeLimit": limit?.timeLimit,
-      "openTime": limit?.openLimit,
-      "apps": limit?.appsTokens.count,
-      "weekdays": limit?.weekdays,
-      "impulseTime": limit?.impulseTime,
-      "usageWarning": limit?.usageWarning
-    ] as [String : Any]
-    
-    resolve(["status": "success", "limit" : limitData])
-  }
-  
   @MainActor @objc
   func updateLimit(
     _ limitId: String,
@@ -616,10 +359,10 @@ class ScreenTimeModule: NSObject {
         return
       }
       // Call disable limit without change status
-      disableLimit(limitId: uuid, updateStore: false)
+      LimitModule.shared.disableLimit(limitId: uuid, updateStore: false)
 
       // Save changes on store
-      let limit = findLimit(limitId: uuid)
+      let limit = LimitModule.shared.findLimit(limitId: uuid)
       limit?.name = name
       limit?.timeLimit = timeLimit
       limit?.openLimit = openLimit
@@ -652,194 +395,24 @@ class ScreenTimeModule: NSObject {
       }
       
       // Enable limit again with the last changes
-      enableLimit(limitId: uuid, updateStore: false)
+      LimitModule.shared.enableLimit(limitId: uuid, updateStore: false)
       resolve(["status": "success"])
     } catch {
       logger.error("Error trying to update limit")
     }
   }
   
-  @objc
-  func readLastLog(_ resolve: @escaping RCTPromiseResolveBlock, rejecter reject: @escaping RCTPromiseRejectBlock) {
-
-    if let lastActivityLog = sharedDefaults?.string(forKey: "lastActivityLog") {
-        print("Última actividad registrada: \(lastActivityLog)")
-    } else {
-        print("No se encontró ningún valor para 'lastActivityLog'")
-    }
-  }
-  
   @MainActor @objc
-  func getBlocks(_ resolve: @escaping RCTPromiseResolveBlock, rejecter reject: @escaping RCTPromiseRejectBlock) {
-    do {
-      let context = try getContext()
-      
-      let fetchDescriptor = FetchDescriptor<Block>()
-      let blocks = try context.fetch(fetchDescriptor)
-      
-      let blocksArray = blocks.map { block -> [String: Any] in
-        return [
-            "id": block.id.uuidString, // Asegúrate de que 'id' sea un UUID
-            "title": block.name, // Reemplaza con los campos de tu modelo
-            "subtitle": "\(block.startTime)-\(block.endTime)",
-            "apps": block.appsTokens.count,
-            "sites": block.webDomainTokens.count,
-            "weekdays": block.weekdays,
-            "enable": block.enable
-        ]
-      }
-      resolve(["status": "success", "blocks" : blocksArray])
-    } catch {
-      print("Error getting blocks")
-    }
-  }
-  
-  @MainActor @objc
-  func deleteBlock(_ blockId: String, resolve: @escaping RCTPromiseResolveBlock, rejecter reject: @escaping RCTPromiseRejectBlock) {
-    do {
-      guard let uuid = UUID(uuidString: blockId) else {
-        reject("invalid_uuid", "El blockId proporcionado no es un UUID válido.", nil)
-        return
-      }
-  
-      let context = try getContext()
-      
-      // Stop monitoring
-      var fetchDescriptor = FetchDescriptor<Block>(
-        predicate: #Predicate{ $0.id == uuid }
-      )
-      fetchDescriptor.fetchLimit = 1
-      let result = try context.fetch(fetchDescriptor)
-      let block = result.first
-      
-      let deviceActivityCenter = DeviceActivityCenter();
-      
-      if block?.weekdays.count == 0 {
-        deviceActivityCenter.stopMonitoring([DeviceActivityName(rawValue: blockId)])
-      } else {
-        let deviceActivityNames: [DeviceActivityName] = block?.weekdays.map { weekday in DeviceActivityName(rawValue: "\(blockId)-day-\(weekday)") } ?? []
-        deviceActivityCenter.stopMonitoring(deviceActivityNames)
-        print(deviceActivityNames)
-      }
-      
-      // Remove restriction
-      let store = ManagedSettingsStore(named: ManagedSettingsStore.Name(rawValue: blockId))
-      store.shield.applications = nil
-      store.shield.webDomains = nil
-      
-      // Delete from store
-      try context.delete(model: Block.self, where: #Predicate { $0.id == uuid })
-      resolve("Block deleted")
-    } catch {
-      reject("Error", "Could not delete block", nil)
-    }
-  }
-  
-  @MainActor @objc
-  func getBlock(_ blockId: String, resolve: @escaping RCTPromiseResolveBlock, rejecter reject: @escaping RCTPromiseRejectBlock) {
-    do {
-      guard let uuid = UUID(uuidString: blockId) else {
-        reject("invalid_uuid", "El blockId proporcionado no es un UUID válido.", nil)
-        return
-      }
-      
-      let context = try getContext()
-      
-      var fetchDescriptor = FetchDescriptor<Block>(
-        predicate: #Predicate{ $0.id == uuid }
-      )
-      fetchDescriptor.fetchLimit = 1
-      let result = try context.fetch(fetchDescriptor)
-      let block = result.first
-      
-      let blockData = [
-        "id": block?.id.uuidString,
-        "name": block?.name,
-        "startTime": block?.startTime,
-        "endTime": block?.endTime,
-        "apps": block?.appsTokens.count,
-        "sites": block?.webDomainTokens.count,
-        "weekdays": block?.weekdays
-      ] as [String : Any]
-
-      resolve(["status": "success", "block" : blockData])
-    } catch {
-      reject("Error", "Could not delete block", nil)
-    }
-  }
-  
-  @MainActor @objc
-  func updateBlockStatus(_ blockId: String, isEnable: Bool, resolve: @escaping RCTPromiseResolveBlock, rejecter reject: @escaping RCTPromiseRejectBlock) {
-    do {
-      guard let uuid = UUID(uuidString: blockId) else {
-        reject("invalid_uuid", "El blockId proporcionado no es un UUID válido.", nil)
-        return
-      }
-      let deviceActivityCenter = DeviceActivityCenter();
-      
-      let context = try getContext()
-      
-      let fetchDescriptor = FetchDescriptor<Block>(
-        predicate: #Predicate { $0.id == uuid }
-      )
-      let result = try context.fetch(fetchDescriptor)
-      let block = result.first
-
-      block?.enable = isEnable
-      
-      if !isEnable {
-        // Stop monitoring
-        if block?.weekdays.count == 0 {
-          try deviceActivityCenter.stopMonitoring([DeviceActivityName(rawValue: blockId)])
-        } else {
-          let deviceActivityNames: [DeviceActivityName] = block?.weekdays.map { weekday in DeviceActivityName(rawValue: "\(blockId)-day-\(weekday)") } ?? []
-          try deviceActivityCenter.stopMonitoring(deviceActivityNames)
-          print(deviceActivityNames)
-        }
-        
-        let store = ManagedSettingsStore(named: ManagedSettingsStore.Name(rawValue: blockId))
-        store.shield.applications = nil
-        store.shield.webDomains = nil
-      } else {
-        let startTimeComponents = block?.startTime.split(separator: ":") ?? []
-        let endTimeComponents = block?.endTime.split(separator: ":") ?? []
-        let weekdays = block?.weekdays ?? []
-        
-        if weekdays.count == 0 {
-          try deviceActivityCenter.startMonitoring(
-            DeviceActivityName(rawValue: blockId),
-            during: DeviceActivitySchedule(
-              intervalStart: DateComponents(hour: Int(startTimeComponents[0]), minute: Int(startTimeComponents[1])),
-              intervalEnd: DateComponents(hour: Int(endTimeComponents[0]), minute: Int(endTimeComponents[1])),
-              repeats: false
-            )
-          )
-          print("Only one time \(blockId)")
-        } else {
-          for weekday in weekdays {
-            try deviceActivityCenter.startMonitoring(
-              DeviceActivityName(rawValue: "\(blockId)-day-\(weekday)"),
-              during: DeviceActivitySchedule(
-                intervalStart: DateComponents(hour: Int(startTimeComponents[0]), minute: Int(startTimeComponents[1]), weekday: weekday),
-                intervalEnd: DateComponents(hour: Int(endTimeComponents[0]), minute: Int(endTimeComponents[1]), weekday: weekday),
-                repeats: true
-              )
-            )
-            print("Repeat on \(weekday) \(blockId)")
-          }
-        }
-      }
-      
-      try context.save()
-
-      resolve(["status": "success", "blockId" : blockId, "isEnable": isEnable, "blockName": block?.name])
-    } catch {
-      print("Error updating block status")
-    }
-  }
-  
-  @MainActor @objc
-  func updateBlock(_ blockId: String, name: String, startTime: String, endTime: String, weekdays: [Int], changeApps: Bool, resolve: @escaping RCTPromiseResolveBlock, rejecter reject: @escaping RCTPromiseRejectBlock) {
+  func updateBlock(
+    _ blockId: String,
+    name: String,
+    startTime: String,
+    endTime: String,
+    weekdays: [Int],
+    changeApps: Bool,
+    resolve: @escaping RCTPromiseResolveBlock,
+    rejecter reject: @escaping RCTPromiseRejectBlock
+  ) {
     do {
       guard let uuid = UUID(uuidString: blockId) else {
         reject("invalid_uuid", "El blockId proporcionado no es un UUID válido.", nil)
